move beta tester form into own component
